ethereum_wallet_tracker
The Ethereum Wallet Tracker is a Python program that allows you to retrieve the value of an Ethereum wallet in US dollars by simply entering the wallet's address. This program makes use of the CoinMarketCap API to fetch the current price of Ethereum and ERC20 tokens and web3 package to get the balance of any wallet.